+++ lookups.py
